move ScalarLikelihood subclasses to separate files
